extract cors to resource server file
