+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-etc-node/client-extension.yaml
@@ -11,3 +11,18 @@ liferay-sample-node-oauth-application-user-agent:
     scopes:
         - Liferay.Headless.Admin.Workflow.everything
     type: oAuthApplicationUserAgent
+liferay-sample-node-object-action-1:
+    name: Liferay Sample Node Object Action 1
+    oAuth2ApplicationExternalReferenceCode: liferay-sample-node-oauth-application-user-agent
+    resourcePath: /sample/object/action/1
+    type: objectAction
+liferay-sample-node-object-action-2:
+    name: Liferay Sample Node Object Action 2
+    oAuth2ApplicationExternalReferenceCode: liferay-sample-node-oauth-application-user-agent
+    resourcePath: /sample/object/action/2
+    type: objectAction
+liferay-sample-node-workflow-action-1:
+    name: Liferay Sample Node Workflow Action 1
+    oAuth2ApplicationExternalReferenceCode: liferay-sample-node-oauth-application-user-agent
+    resourcePath: /sample/workflow/action/1
+    type: workflowAction
