+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-etc-spring-boot/src/main/resources/application-default.properties
@@ -10,7 +10,7 @@ com.liferay.lxc.dxp.server.protocol=http
 # OAuth
 #
 
-liferay.oauth.application.external.reference.codes=liferay-sample-oauth-application-user-agent
+liferay.oauth.application.external.reference.codes=liferay-sample-sample-spring-oauth-application-user-agent
 
 #
 # Spring Boot
